Soriya category

See merge request Chanthai1102/only-admin!24

+++ src/main/java/com/kiloit/onlyadmin/database/entity/PostEntity.java
@@ -16,7 +16,7 @@ public class PostEntity extends BaseEntity {
     private String thumbnail;
     private Boolean status;
     @Column(name = "media_id")
-    private Integer listMediaId;
+    private Long listMediaId;
 
     @Column(name = "public_at")
     private Instant publicAt;
